// src/models/button-mode.ts
export const enum ButtonMode {
  AddFunction = 1,
  EditFunction = 2,
  DeleteFunction = 3,
  AddBlock = 4,
  EditBlock = 5,
  DeleteBlock = 6,
  AddRegister = 7,
  EditRegister = 8,
  DeleteRegister = 9,
  AddField = 10,
  EditField = 11,
  DeleteField = 12,
  AddEnumeratedValue = 13,
  EditEnumeratedValue = 14,
  DeleteEnumeratedValue = 15,
  EditBits = 16,
  Any = 17,
  DeleteOpenedRegister = 18
}